real-time-marketing-pipeline/airflow/requirements.txt
boto3
pandas
psycopg2-binary
apache-airflow==2.8.1
